Split out opaque from type_of
